fix: move NotificationBell to core and add Notifications row to ProfileScreen
